Refactor and cleanup
* refactors a number of "waitForIncidentAndX" to a generic function handler
* splits related code into a number of different files for organizational purposes
* adds "open in browser" feature; currently hardcoded in xdg-open
