+++ pkg/tui/tui.go
@@ -2,10 +2,10 @@ package tui
 
 import (
 	"fmt"
+	"strings"
 	"time"
 
 	"github.com/PagerDuty/go-pagerduty"
-	"github.com/charmbracelet/bubbles/key"
 	"github.com/charmbracelet/bubbles/table"
 	tea "github.com/charmbracelet/bubbletea"
 )
@@ -13,6 +13,8 @@ import (
 const (
 	waitTime           = time.Millisecond * 1
 	defaultInputPrompt = " $ "
+
+	nilNoteErr = "incident note content is empty"
 )
 
 type errMsg struct{ error }
@@ -32,52 +34,13 @@ func (m model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 	switch msg := msg.(type) {
 
 	case errMsg:
-		debug(fmt.Sprintf("errMsg: %v", msg))
-		m.setStatus(msg.Error())
-		m.err = msg
-		return m, nil
+		return m.errMsgHandler(msg)
 
 	case tea.WindowSizeMsg:
-		debug("tea.WindowSizeMsg")
-		windowSize = msg
-		top, _, bottom, _ := mainStyle.GetMargin()
-		eighthWindow := windowSize.Width / 8
-		cellPadding := (horizontalPadding * 2) * 4
-		borderEdges := 2 + 10
-
-		m.help.Width = windowSize.Width - borderEdges
-
-		m.table.SetColumns([]table.Column{
-			{Title: dot, Width: 1},
-			{Title: "ID", Width: eighthWindow + cellPadding - borderEdges},
-			{Title: "Summary", Width: eighthWindow * 3},
-			{Title: "Service", Width: eighthWindow * 3},
-		})
-
-		height := windowSize.Height - top - bottom - 10
-
-		m.table.SetHeight(height)
-
-		m.incidentViewer.Width = windowSize.Width - borderEdges
-		m.incidentViewer.Height = height
+		return m.windowSizeMsgHandler(msg)
 
 	case tea.KeyMsg:
-		debug("tea.KeyMsg", fmt.Sprint(msg))
-		if key.Matches(msg, defaultKeyMap.Quit) {
-			return m, tea.Quit
-		}
-
-		// Default commands for the table view
-		switch {
-		case m.err != nil:
-			return switchErrorFocusMode(m, msg)
-		case m.viewingIncident:
-			return switchIncidentFocusMode(m, msg)
-		case m.input.Focused():
-			return switchInputFocusMode(m, msg)
-		case m.table.Focused():
-			return switchTableFocusMode(m, msg)
-		}
+		return m.keyMsgHandler(msg)
 
 	// Command to get an incident by ID
 	case getIncidentMsg:
@@ -87,12 +50,11 @@ func (m model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 
 	// Set the selected incident to the incident returned from the getIncident command
 	case gotIncidentMsg:
-		debug("gotIncidentMsg", "TRUNCATED")
 		if msg.err != nil {
-			return m, func() tea.Msg {
-				return errMsg{msg.err}
-			}
+			return m, func() tea.Msg { return errMsg{msg.err} }
 		}
+
+		debug("gotIncidentMsg", msg.incident.ID)
 
 		m.setStatus(fmt.Sprintf("got incident %s", msg.incident.ID))
 		m.selectedIncident = msg.incident
@@ -102,12 +64,19 @@ func (m model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 		)
 
 	case gotIncidentNotesMsg:
-		debug("gotIncidentNotesMsg", "TRUNCATED")
 		if msg.err != nil {
-			return m, func() tea.Msg {
-				return errMsg{msg.err}
-			}
+			return m, func() tea.Msg { return errMsg{msg.err} }
 		}
+
+		getIDs := func(x []pagerduty.IncidentNote) []string {
+			var ids []string
+			for _, n := range x {
+				ids = append(ids, n.ID)
+			}
+			return ids
+		}
+
+		debug("gotIncidentNotesMsg", strings.Join(getIDs(msg.notes), ", "))
 
 		// CANNOT refer to the m.SelectedIncident, because it may not have
 		// completed yet, and will be nil
@@ -124,12 +93,19 @@ func (m model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 		}
 
 	case gotIncidentAlertsMsg:
-		debug("gotIncidentAlertsMsg", "TRUNCATED")
 		if msg.err != nil {
-			return m, func() tea.Msg {
-				return errMsg{msg.err}
-			}
+			return m, func() tea.Msg { return errMsg{msg.err} }
 		}
+
+		getIDs := func(x []pagerduty.IncidentAlert) []string {
+			var ids []string
+			for _, n := range x {
+				ids = append(ids, n.ID)
+			}
+			return ids
+		}
+
+		debug("gotIncidentAlertsMsg", strings.Join(getIDs(msg.alerts), ", "))
 
 		// CANNOT refer to the m.SelectedIncident, because it may not have
 		// completed yet, and will be nil
@@ -152,12 +128,19 @@ func (m model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 		cmds = append(cmds, updateIncidentList(m.config))
 
 	case updatedIncidentListMsg:
-		debug("updatedIncidentListMsg", "TRUNCATED")
 		if msg.err != nil {
-			return m, func() tea.Msg {
-				return errMsg{msg.err}
-			}
+			return m, func() tea.Msg { return errMsg{msg.err} }
 		}
+
+		getIDs := func(x []pagerduty.Incident) []string {
+			var ids []string
+			for _, n := range x {
+				ids = append(ids, n.ID)
+			}
+			return ids
+		}
+
+		debug("updatedIncidentListMsg", strings.Join(getIDs(msg.incidents), ", "))
 
 		m.incidentList = msg.incidents
 
@@ -184,12 +167,11 @@ func (m model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 		}
 
 	case editorFinishedMsg:
-		debug("editorFinishedMsg", fmt.Sprint(msg))
 		if msg.err != nil {
-			return m, func() tea.Msg {
-				return errMsg{msg.err}
-			}
+			return m, func() tea.Msg { return errMsg{msg.err} }
 		}
+
+		debug("editorFinishedMsg", fmt.Sprint(msg))
 
 		if m.selectedIncident == nil {
 			m.setStatus("failed to add note - no selected incident")
@@ -200,6 +182,14 @@ func (m model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 
 	// Refresh the local copy of the incident after the note is added
 	case addedIncidentNoteMsg:
+		if msg.err != nil {
+			if msg.err.Error() == nilNoteErr {
+				m.status = "skipping adding empty note to incident"
+				return m, nil
+			}
+			return m, func() tea.Msg { return errMsg{msg.err} }
+		}
+
 		debug("addedIncidentNoteMsg", fmt.Sprint(msg))
 
 		if m.selectedIncident == nil {
@@ -226,9 +216,28 @@ func (m model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 		m.setStatus(fmt.Sprintf("multiple alerts for incident - logging into cluster %s from first alert %s", cluster, m.selectedIncidentAlerts[0].ID))
 		return m, func() tea.Msg { return login(cluster, m.launcher) }
 
+	case loginFinishedMsg:
+		if msg.err != nil {
+			return m, func() tea.Msg { return errMsg{msg.err} }
+		}
+
+	case openBrowserMsg:
+		debug("openBrowserMsg", fmt.Sprint(msg))
+		return m, func() tea.Msg { return openBrowserCmd([]string{"xdg-open"}, m.selectedIncident.HTMLURL) }
+
+	case browserFinishedMsg:
+		if msg.err != nil {
+			return m, func() tea.Msg { return errMsg{msg.err} }
+		} else {
+			m.setStatus(fmt.Sprintf("opened incident %s in browser - check browser window", m.selectedIncident.ID))
+			return m, nil
+		}
+
+	// This is a catch all for any action that requires a selected incident
+	//
 	case waitForSelectedIncidentThenDoMsg:
-		debug("waitForSelectedIncidentThenDoMsg", fmt.Sprint(msg.action, msg.msg))
-		if msg.action == "" {
+		debug("waitForSelectedIncidentThenDoMsg: ", fmt.Sprintf("%+v, %+v", msg.action, msg.msg))
+		if msg.action == nil {
 			m.setStatus("failed to perform action: no action included in msg")
 			return m, nil
 		}
@@ -240,36 +249,11 @@ func (m model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 		if m.selectedIncident == nil {
 			time.Sleep(waitTime)
 			m.setStatus("waiting for incident info...")
-			return m, func() tea.Msg { return waitForSelectedIncidentThenDoMsg(msg) }
+			return m, func() tea.Msg { return waitForSelectedIncidentThenDoMsg{action: msg.action, msg: msg.msg} }
 		}
 
-		// TODO https://github.com/clcollins/srepd/issues/2: Figure out how to use an interface for the msg.action to write this once
-		// cmds = append(cmds, func() tea.Msg { return msg.action(msg.msg) })
-
-		switch msg.action {
-		// TODO https://github.com/clcollins/srepd/issues/2: See TODO above
-		// case "acknowledgeIncidentsMsg":
-		// 	if msg.msg.incidents == nil {
-		// 		m.setStatus("failed acknowledging incidents - no incidents provided")
-		// 		return m, nil
-		// 	}
-
-		// 	return m, tea.Sequence(
-		// 		acknowledgeIncidents(m.config, msg.incidents),
-		// 		func() tea.Msg { return clearSelectedIncidentsMsg("sender: acknowledgeIncidentsMsg") },
-		// 	)
-		// case "annotateIncidentsMsg":
-		// TODO: The "ACTION" here needs to be a tea.Msg, not a string
-		case "loginMsg":
-			return m, func() tea.Msg { return loginMsg("login") }
-		// case "reassignIncidentsMsg":
-		case "renderIncidentMsg":
-			return m, func() tea.Msg { return renderIncidentMsg("render") }
-		// case "silenceIncidentsMsg":
-		default:
-			debug(fmt.Sprintf("%v not implemented", msg.action))
-			return m, nil
-		}
+		debug("Trying to do action")
+		cmds = append(cmds, msg.action)
 
 	case renderIncidentMsg:
 		debug("renderIncidentMsg", fmt.Sprint(msg))
@@ -281,21 +265,12 @@ func (m model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 		cmds = append(cmds, renderIncident(&m))
 
 	case renderedIncidentMsg:
-		debug("renderedIncidentMsg", fmt.Sprint(msg))
-		// TODO https://github.com/clcollins/srepd/issues/3 - check the msg.err properly
-		// not in the renderIncident() function
+		if msg.err != nil {
+			return m, func() tea.Msg { return errMsg{msg.err} }
+		}
+		debug("renderedIncidentMsg")
 		m.incidentViewer.SetContent(msg.content)
 		m.viewingIncident = true
-
-	case waitForSelectedIncidentsThenAnnotateMsg:
-		debug("waitForSelectedIncidentsThenAnnotateMsg", fmt.Sprint(msg))
-		if m.selectedIncident == nil {
-			time.Sleep(waitTime)
-			m.setStatus("waiting for incident info...")
-			return m, func() tea.Msg { return waitForSelectedIncidentsThenAnnotateMsg(msg) }
-		}
-		// TODO - https://github.com/clcollins/srepd/issues/2: This needs to be a tea.Cmd to allow waitForSelectedIncidentsThenDo
-		cmds = append(cmds, openEditorCmd(m.editor))
 
 	case acknowledgeIncidentsMsg:
 		debug("acknowledgeIncidentsMsg", fmt.Sprint(msg))
@@ -310,6 +285,9 @@ func (m model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 		)
 
 	case acknowledgedIncidentsMsg:
+		if msg.err != nil {
+			return m, func() tea.Msg { return errMsg{msg.err} }
+		}
 		debug("acknowledgedIncidentsMsg", fmt.Sprint(msg))
 		m.setStatus(fmt.Sprintf("acknowledged incidents %v; refreshing Incident List ", msg))
 		return m, func() tea.Msg { return updateIncidentListMsg("sender: acknowledgedIncidentsMsg") }
@@ -358,17 +336,6 @@ func (m model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 			func() tea.Msg { return clearSelectedIncidentsMsg("sender: silenceIncidentsMsg") },
 		)
 
-	// There is no "silencedIncidentsMsg" - silences are really reassignments under the hood
-
-	case waitForSelectedIncidentsThenSilenceMsg:
-		debug("waitForSelectedIncidentsThenSilenceMsg", fmt.Sprint(msg))
-		if m.selectedIncident == nil {
-			time.Sleep(waitTime)
-			m.setStatus("waiting for incident info...")
-			return m, func() tea.Msg { return waitForSelectedIncidentsThenSilenceMsg(msg) }
-		}
-		return m, func() tea.Msg { return silenceIncidentsMsg{incidents: []*pagerduty.Incident{m.selectedIncident}} }
-
 	case clearSelectedIncidentsMsg:
 		debug("clearSelectedIncidentsMsg", fmt.Sprint(msg))
 		m.viewingIncident = false
@@ -380,185 +347,4 @@ func (m model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 
 	return m, tea.Batch(cmds...)
 
-}
-
-func (m model) View() string {
-	debug("View")
-	helpView := helpStyle.Render(m.help.View(defaultKeyMap))
-
-	switch {
-	case m.err != nil:
-		debug("error")
-		return (errorStyle.Render(dot+"ERROR"+dot+"\n\n"+m.err.Error()) + "\n" + helpView)
-
-	case m.viewingIncident:
-		debug("viewingIncident")
-		return mainStyle.Render(m.renderHeader() + "\n" + m.incidentViewer.View() + "\n" + helpView)
-	default:
-		tableView := tableContainerStyle.Render(m.table.View())
-		if m.input.Focused() {
-			debug("viewingTable and input")
-			return mainStyle.Render(m.renderHeader() + "\n" + tableView + "\n" + m.input.View() + "\n" + helpView)
-		}
-		debug("viewingTable")
-		return mainStyle.Render(m.renderHeader() + "\n" + tableView + "\n" + helpView)
-	}
-}
-
-// tableFocusMode is the main mode for the application
-func switchTableFocusMode(m model, msg tea.Msg) (tea.Model, tea.Cmd) {
-	debug("switchTableFocusMode")
-	var cmds []tea.Cmd
-
-	switch msg := msg.(type) {
-	case tea.KeyMsg:
-		switch {
-		case key.Matches(msg, defaultKeyMap.Help):
-			m.help.ShowAll = !m.help.ShowAll
-
-		case key.Matches(msg, defaultKeyMap.Up):
-			m.table.MoveUp(1)
-
-		case key.Matches(msg, defaultKeyMap.Down):
-			m.table.MoveDown(1)
-
-		case key.Matches(msg, defaultKeyMap.Top):
-			m.table.GotoTop()
-
-		case key.Matches(msg, defaultKeyMap.Bottom):
-			m.table.GotoBottom()
-
-		case key.Matches(msg, defaultKeyMap.Enter):
-			m.viewingIncident = true
-			return m, tea.Sequence(
-				func() tea.Msg { return getIncidentMsg(m.table.SelectedRow()[1]) },
-				func() tea.Msg { return waitForSelectedIncidentThenDoMsg{action: "renderIncidentMsg", msg: "render"} },
-			)
-
-		case key.Matches(msg, defaultKeyMap.Team):
-			m.teamMode = !m.teamMode
-			cmds = append(cmds, func() tea.Msg { return updatedIncidentListMsg{m.incidentList, nil} })
-
-		case key.Matches(msg, defaultKeyMap.Refresh):
-			m.setStatus(loadingIncidentsStatus)
-			cmds = append(cmds, updateIncidentList(m.config))
-
-		// In table mode, highlighted incidents are not selected yet, so they need to be retrieved
-		// and then can be acted upon.  Since tea.Sequence does not wait for completion, the
-		// "waitForSelectedIncidentsThen..." functions are used to wait for the selected incident
-		// to be retrieved from PagerDuty
-		case key.Matches(msg, defaultKeyMap.Silence):
-			return m, tea.Sequence(
-				func() tea.Msg { return getIncidentMsg(m.table.SelectedRow()[1]) },
-				func() tea.Msg { return waitForSelectedIncidentsThenSilenceMsg("wait") },
-			)
-
-		case key.Matches(msg, defaultKeyMap.Ack):
-			return m, tea.Sequence(
-				func() tea.Msg { return getIncidentMsg(m.table.SelectedRow()[1]) },
-				func() tea.Msg { return waitForSelectedIncidentsThenAcknowledgeMsg("wait") },
-			)
-
-		case key.Matches(msg, defaultKeyMap.Note):
-			return m, tea.Sequence(
-				func() tea.Msg { return getIncidentMsg(m.table.SelectedRow()[1]) },
-				func() tea.Msg { return waitForSelectedIncidentsThenAnnotateMsg("wait") },
-			)
-
-		case key.Matches(msg, defaultKeyMap.Login):
-			return m, tea.Sequence(
-				func() tea.Msg { return getIncidentMsg(m.table.SelectedRow()[1]) },
-				func() tea.Msg { return waitForSelectedIncidentThenDoMsg{action: "loginMsg", msg: "wait"} },
-			)
-
-		case key.Matches(msg, defaultKeyMap.Input):
-			return m, tea.Sequence(
-				m.input.Focus(),
-			)
-		}
-	}
-	return m, tea.Batch(cmds...)
-}
-
-func switchInputFocusMode(m model, msg tea.Msg) (tea.Model, tea.Cmd) {
-	debug("switchInputFocusMode")
-	var cmds []tea.Cmd
-
-	switch msg := msg.(type) {
-	case tea.KeyMsg:
-		switch {
-		case key.Matches(msg, defaultKeyMap.Help):
-			m.help.ShowAll = !m.help.ShowAll
-
-		case key.Matches(msg, defaultKeyMap.Back):
-			m.input.Blur()
-			m.table.Focus()
-			m.input.Prompt = defaultInputPrompt
-			return m, nil
-
-		case key.Matches(msg, defaultKeyMap.Enter):
-
-		}
-	}
-	return m, tea.Batch(cmds...)
-}
-
-func switchIncidentFocusMode(m model, msg tea.Msg) (tea.Model, tea.Cmd) {
-	debug("switchIncidentFocusMode")
-	var cmd tea.Cmd
-	var cmds []tea.Cmd
-
-	switch msg := msg.(type) {
-	case tea.KeyMsg:
-		switch {
-		case key.Matches(msg, defaultKeyMap.Help):
-			m.help.ShowAll = !m.help.ShowAll
-
-		// This un-sets the selected incident and returns to the table view
-		case key.Matches(msg, defaultKeyMap.Back):
-			m.viewingIncident = false
-			m.selectedIncident = nil
-			m.selectedIncidentAlerts = nil
-			m.selectedIncidentNotes = nil
-
-		case key.Matches(msg, defaultKeyMap.Ack):
-			return m, func() tea.Msg { return acknowledgeIncidentsMsg{incidents: []*pagerduty.Incident{m.selectedIncident}} }
-
-		case key.Matches(msg, defaultKeyMap.Silence):
-			return m, func() tea.Msg { return silenceIncidentsMsg{incidents: []*pagerduty.Incident{m.selectedIncident}} }
-
-		case key.Matches(msg, defaultKeyMap.Note):
-			cmds = append(cmds, openEditorCmd(m.editor))
-
-		case key.Matches(msg, defaultKeyMap.Refresh):
-			return m, func() tea.Msg { return getIncidentMsg(m.selectedIncident.ID) }
-
-		case key.Matches(msg, defaultKeyMap.Login):
-			return m, tea.Sequence(
-				func() tea.Msg { return getIncidentMsg(m.table.SelectedRow()[1]) },
-				func() tea.Msg { return waitForSelectedIncidentThenDoMsg{action: "loginMsg", msg: "wait"} },
-			)
-		}
-	}
-
-	m.incidentViewer, cmd = m.incidentViewer.Update(msg)
-	cmds = append(cmds, cmd)
-
-	return m, tea.Batch(cmds...)
-}
-
-func switchErrorFocusMode(m model, msg tea.Msg) (tea.Model, tea.Cmd) {
-	debug("switchErrorFocusMode")
-	switch msg := msg.(type) {
-	case tea.KeyMsg:
-		switch {
-		case key.Matches(msg, defaultKeyMap.Help):
-			m.help.ShowAll = !m.help.ShowAll
-
-		case key.Matches(msg, defaultKeyMap.Back):
-			m.err = nil
-			m.setStatus("")
-		}
-	}
-	return m, nil
 }
